user.enroll: add icons next to input fields, align header to center
Right now the only problem is that input boxes don't have the same
width. I tried bunch of solutions for it, yet I'm unable to fix it.
If I somehow remove the parent `<div>` of the `<input>`, it kinda
works, but the leftside of the box becomes straight, not-rounded.
Also I couldn't figure out how to remove the parent `<div>`, crispy
adds it automatically.

Opened a stackoverflow question for this problem.

Related to #1