user.enroll: remove crispy, localize error messages
I haven't mentioned about this in the previous commit, but one
"weakness" of this implementation is that the error messages are
not dynamic. When do you need dynamic error messages? For
max_length constraints for example. I haven't look out solutions,
leaving as it is right now. So, I kindly remind you that if you
plat with the constraints of a model -in this case, it's member-
you should also change the error messages.

P.S. Dynamic error messages exist by default in Django.

One other thing I haven't mentioned is the new ifs at the clean()
method of forms. When is_valid() called, apparently Django first
calls clean method instead of checking for the required fields
first. So, I had to fix an unmentioned bug by checking whether the
field exists or not before trying to clean it.